#57 [FEAT] 신고 게시물 api 로직 작성

+++ app/src/main/java/com/kpaas/plog/presentation/report/navigation/ReportNavigator.kt
@@ -13,7 +13,7 @@ class ReportNavigator(
         }
     }
 
-    fun navigateReportContent(id: Int) {
+    fun navigateReportContent(id: Long) {
         val route = "reportContent?id=${id}"
         navController.navigate(route) {
             popUpTo(navController.graph.startDestinationId) {
